newer sydtest

+++ smos-server-gen/src/Smos/Server/TestUtils.hs
@@ -94,7 +94,8 @@ serverSetupFunc' man = wrapSetupFunc $ \pool -> do
             }
     pure $ Server.makeSyncApp env
   p <- unwrapSetupFunc applicationSetupFunc application
-  pure $ mkClientEnv man (BaseUrl Http "127.0.0.1" p "")
+  -- The fromIntegral is safe because it's PortNumber -> Int
+  pure $ mkClientEnv man (BaseUrl Http "127.0.0.1" (fromIntegral p) "")
 
 registerLogin :: Register -> Login
 registerLogin register =
